all apis are formed along with JWT authentication between the routes

+++ backend/controller/goalController.js
@@ -1,12 +1,15 @@
 const asyncHandler = require("express-async-handler")
 const Goal = require('../models/goalModel')
+const User = require("../models/userModel")
 
 // desc     Get Goals
 // route    GET /api/goals
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
-    res.status(200).json(goals)
+    res.status(200).json({
+        goals
+    })
 })
 
 
@@ -19,6 +22,7 @@ const setGoals = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text,
     })
     res.status(200).json(goal)
@@ -36,6 +40,18 @@ const updateGoals = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error("user not found");
+    }
+
+    // make sure the logged in user matches the goal user
+    if (user.id !== goal.user.toString()) {
+        res.status(401);
+        throw new Error("User not authorized")
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -53,6 +69,19 @@ const deleteGoals = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error("user not found");
+    }
+
+    // make sure the logged in user matches the goal user
+    if (user.id !== goal.user.toString()) {
+        res.status(401);
+        throw new Error("User not authorized")
     }
 
     await Goal.deleteOne({ _id: req.params.id })
